lambda/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# lambda/lambda.py: 5
psycopg2_binary == 2.9.1

# lambda/lambda.py: 3
pytz == 2020.4

# lambda/lambda.py: 1
requests == 2.23.0
